pyproject.toml
[project]
name = "open-jet"
version = "0.3.0"
description = "Offline local-LLM terminal app for Jetson and edge Linux: chat with on-device models, run agent tools, and manage context safely."
readme = { file = "README.pypi.md", content-type = "text/markdown" }
requires-python = ">=3.10"
authors = [
    { name = "Louis" }
]
keywords = ["llm", "local-llm", "ai-agent", "tui", "jetson", "edge-ai", "offline", "llama.cpp"]
license = "AGPL-3.0-only"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: User Interfaces",
    "Typing :: Typed",
]
dependencies = [
    "faster-whisper>=1.2",
    "httpx>=0.27",
    "opentelemetry-exporter-otlp-proto-http>=1.40",
    "opentelemetry-sdk>=1.40",
    "prompt_toolkit>=3.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tiktoken>=0.7",
]

[project.urls]
Homepage = "https://openjet.dev/"
Website = "https://openjet.dev/"
Repository = "https://github.com/l-forster/open-jet"
Issues = "https://github.com/l-forster/open-jet/issues"

[project.scripts]
open-jet = "src.cli:main"
openjet = "src.cli:main"

[build-system]
requires = ["setuptools>=68,<80", "wheel", "Cython>=3.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = false
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
include = ["src", "src.*", "open_jet", "open_jet.*"]

[tool.setuptools.package-data]
src = ["app.tcss"]
